refactor: 라디오 컴포넌트 props 주석 추가
