+++ src/components/@common/Form/Radio/index.tsx
@@ -3,13 +3,17 @@ import * as Styled from './Radio.styled';
 
 export interface IRadioProps
   extends React.InputHTMLAttributes<HTMLInputElement> {
-  text?: string;
+  /** 라디오 버튼과 연결된 라벨 */
+  label: string;
+  /** 라디오 버튼의 고유 ID */
   id?: string;
+  /** 폼에서 사용할 필드 이름 */
   name: string;
+  /** react-hook-form의 유효성 검사 규칙 */
   rules?: RegisterOptions;
 }
 
-const Radio = ({ text, id, name, rules, ...rest }: IRadioProps) => {
+const Radio = ({ label, id, name, rules, ...rest }: IRadioProps) => {
   const { register } = useFormContext();
 
   return (
@@ -22,7 +26,7 @@ const Radio = ({ text, id, name, rules, ...rest }: IRadioProps) => {
           {...rest}
         />
         <Styled.CheckMark title="체크" width="16" height="16" />
-        {text}
+        {label}
       </Styled.Label>
     </>
   );
