+++ src/components/@common/ImageViewer/index.tsx
@@ -11,9 +11,13 @@ import useLockBodyScroll from '@/hooks/@common/useLockBodyScroll';
 import PageNavigator from '@/components/home/PageNavigator';
 
 interface IImageViewer {
+  /** 표시할 이미지 목록 (이미지 경로 배열) */
   imageList: string[];
+  /** 이미지의 대체 텍스트 */
   alt: string;
+  /** 모달을 닫을 때 실행할 함수 */
   onClose: () => void;
+  /** 초기 선택된 이미지의 인덱스 */
   selectedImageIndex: number;
 }
